requirements.txt
rich>=13.0.0
requests>=2.28.0
phonenumbers>=8.13.0
python-whois>=0.9.0
dnspython>=2.3.0
Pillow>=10.0.0
